+++ examples/quota_project/main.tf
@@ -26,7 +26,7 @@ resource "random_string" "suffix" {
 
 module "quota-project" {
   source  = "terraform-google-modules/project-factory/google"
-  version = "~> 15.0"
+  version = "~> 16.0"
 
   name              = "pf-ci-test-quota-${random_string.suffix.result}"
   random_project_id = true
@@ -58,4 +58,6 @@ module "quota-project" {
       value      = "95"
     }
   ]
+
+  deletion_policy = "DELETE"
 }