+++ examples/shared_vpc/main.tf
@@ -24,7 +24,7 @@ locals {
  *****************************************/
 module "host-project" {
   source  = "terraform-google-modules/project-factory/google"
-  version = "~> 15.0"
+  version = "~> 16.0"
 
   random_project_id              = true
   name                           = var.host_project_name
@@ -39,6 +39,7 @@ module "host-project" {
     "cloudresourcemanager.googleapis.com"
   ]
 
+  deletion_policy = "DELETE"
 }
 
 /******************************************
@@ -93,7 +94,7 @@ module "vpc" {
  *****************************************/
 module "service-project" {
   source  = "terraform-google-modules/project-factory/google//modules/svpc_service_project"
-  version = "~> 15.0"
+  version = "~> 16.0"
 
   name              = var.service_project_name
   random_project_id = false
@@ -113,6 +114,7 @@ module "service-project" {
   ]
 
   disable_services_on_destroy = false
+  deletion_policy             = "DELETE"
 }
 
 /******************************************
@@ -120,7 +122,7 @@ module "service-project" {
  *****************************************/
 module "service-project-b" {
   source  = "terraform-google-modules/project-factory/google//modules/svpc_service_project"
-  version = "~> 15.0"
+  version = "~> 16.0"
 
   name              = "b-${var.service_project_name}"
   random_project_id = false
@@ -146,6 +148,7 @@ module "service-project-b" {
   }]
 
   disable_services_on_destroy = false
+  deletion_policy             = "DELETE"
 }
 
 /******************************************
@@ -154,7 +157,7 @@ module "service-project-b" {
  *****************************************/
 module "service-project-c" {
   source  = "terraform-google-modules/project-factory/google//modules/svpc_service_project"
-  version = "~> 15.0"
+  version = "~> 16.0"
 
   name              = "c-${var.service_project_name}"
   random_project_id = false
@@ -184,6 +187,7 @@ module "service-project-c" {
 
   disable_services_on_destroy = false
   grant_network_role          = false
+  deletion_policy             = "DELETE"
 }
 
 /******************************************