+++ modules/app_engine/versions.tf
@@ -20,7 +20,7 @@ terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 3.43, < 6"
+      version = ">= 3.43, < 7"
     }
   }
 }