+++ modules/core_project_factory/main.tf
@@ -77,8 +77,15 @@ resource "google_project" "main" {
   folder_id           = local.project_folder_id
   billing_account     = var.billing_account
   auto_create_network = var.auto_create_network
+  deletion_policy     = var.deletion_policy
 
   labels = var.labels
+
+  lifecycle {
+    ignore_changes = [
+      labels["firebase"],
+    ]
+  }
 }
 
 /******************************************