+++ modules/essential_contacts/versions.tf
@@ -19,17 +19,17 @@ terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 3.43, < 6"
+      version = ">= 3.43, < 7"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 3.43, < 6"
+      version = ">= 3.43, < 7"
     }
   }
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:essential_contacts/v15.0.1"
+    module_name = "blueprints/terraform/terraform-google-project-factory:essential_contacts/v16.0.1"
   }
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:essential_contacts/v15.0.1"
+    module_name = "blueprints/terraform/terraform-google-project-factory:essential_contacts/v16.0.1"
   }
 }