+++ modules/fabric-project/versions.tf
@@ -20,13 +20,13 @@ terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 3.43, < 6"
+      version = ">= 5.41, < 7"
     }
   }
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:fabric-project/v15.0.1"
+    module_name = "blueprints/terraform/terraform-google-project-factory:fabric-project/v16.0.1"
   }
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-project-factory:fabric-project/v15.0.1"
+    module_name = "blueprints/terraform/terraform-google-project-factory:fabric-project/v16.0.1"
   }
 }