+++ modules/svpc_service_project/main.tf
@@ -59,6 +59,7 @@ module "project-factory" {
   default_service_account           = var.default_service_account
   disable_dependent_services        = var.disable_dependent_services
   default_network_tier              = var.default_network_tier
+  deletion_policy                   = var.deletion_policy
 }
 
 /******************************************
@@ -69,12 +70,15 @@ module "shared_vpc_access" {
   host_project_id                    = var.shared_vpc
   enable_shared_vpc_service_project  = true
   service_project_id                 = module.project-factory.project_id
-  active_apis                        = module.project-factory.enabled_apis
+  active_apis                        = var.activate_apis
   shared_vpc_subnets                 = var.shared_vpc_subnets
   service_project_number             = module.project-factory.project_number
   lookup_project_numbers             = false
   grant_services_security_admin_role = var.grant_services_security_admin_role
   grant_network_role                 = var.grant_network_role
+  # Workaround for import complaining about count cannot determine resource instances
+  # until apply. https://github.com/hashicorp/terraform/issues/24690
+  depends_on = [module.project-factory.enabled_apis]
 }
 
 /******************************************