+++ versions.tf
@@ -19,17 +19,17 @@ terraform {
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 5.33, < 6"
+      version = ">= 5.41, < 7"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 5.33, < 6"
+      version = ">= 5.41, < 7"
     }
   }
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-project-factory/v15.0.1"
+    module_name = "blueprints/terraform/terraform-google-project-factory/v16.0.1"
   }
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-project-factory/v15.0.1"
+    module_name = "blueprints/terraform/terraform-google-project-factory/v16.0.1"
   }
 }